requirements.txt
aiofiles==23.1.0
anyio==3.6.2
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
motor==3.1.2
orjson==3.8.10
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
stripe==5.4.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
